+++ cmd/starcoin/src/cli_state.rs
@@ -3,12 +3,16 @@
 
 use std::convert::TryInto;
 use std::env::current_dir;
+use std::fs::File;
 use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use std::time::Duration;
 
 use anyhow::{bail, format_err, Result};
 use serde::de::DeserializeOwned;
+use starcoin_crypto::hash::PlainCryptoHash;
+use starcoin_crypto::multi_ed25519::multi_shard::MultiEd25519SignatureShard;
+use starcoin_crypto::multi_ed25519::MultiEd25519PublicKey;
 use starcoin_crypto::HashValue;
 
 use bcs_ext::BCSCodec;
@@ -27,12 +31,11 @@ use starcoin_vm_types::account_address::AccountAddress;
 use starcoin_vm_types::account_config::association_address;
 use starcoin_vm_types::move_resource::MoveResource;
 use starcoin_vm_types::token::stc::STC_TOKEN_CODE_STR;
-use starcoin_vm_types::transaction::authenticator::AccountPublicKey;
+use starcoin_vm_types::transaction::authenticator::{AccountPublicKey, TransactionAuthenticator};
 use starcoin_vm_types::transaction::{
     DryRunTransaction, RawUserTransaction, SignedUserTransaction, TransactionPayload,
 };
 
-use crate::mutlisig_transaction::sign_multisig_txn_to_file;
 use crate::view::{ExecuteResultView, ExecutionOutputView, TransactionOptions};
 
 static G_HISTORY_FILE_NAME: &str = "history";
@@ -301,12 +304,14 @@ impl CliState {
             AccountPublicKey::Multi(m) => m.clone(),
         };
 
-        let _ = sign_multisig_txn_to_file(
+        let _ = self.sign_multisig_txn_to_file_or_submit(
             sender.address,
             multisig_public_key,
             None,
             signed_txn,
             current_dir()?,
+            true,
+            blocking,
         );
 
         Ok(execute_result)
@@ -322,6 +327,80 @@ impl CliState {
 
     pub fn into_inner(self) -> (ChainNetworkID, Arc<RpcClient>, Option<NodeHandle>) {
         (self.net, self.client, self.node_handle)
+    }
+
+    // Sign multisig transaction, if enough signatures collected & submit is true,
+    // try to submit txn directly.
+    // Otherwise, keep signatures into file.
+    pub fn sign_multisig_txn_to_file_or_submit(
+        &self,
+        sender: AccountAddress,
+        multisig_public_key: MultiEd25519PublicKey,
+        existing_signatures: Option<MultiEd25519SignatureShard>,
+        partial_signed_txn: SignedUserTransaction,
+        output_dir: PathBuf,
+        submit: bool,
+        blocking: bool,
+    ) -> Result<PathBuf> {
+        let my_signatures = if let TransactionAuthenticator::MultiEd25519 { signature, .. } =
+            partial_signed_txn.authenticator()
+        {
+            MultiEd25519SignatureShard::new(signature, *multisig_public_key.threshold())
+        } else {
+            unreachable!()
+        };
+
+        // merge my signatures with existing signatures of other participants.
+        let merged_signatures = {
+            let mut signatures = vec![];
+            if let Some(s) = existing_signatures {
+                signatures.push(s);
+            }
+            signatures.push(my_signatures);
+            MultiEd25519SignatureShard::merge(signatures)?
+        };
+        eprintln!(
+            "mutlisig txn(address: {}, threshold: {}): {} signatures collected",
+            sender,
+            merged_signatures.threshold(),
+            merged_signatures.signatures().len()
+        );
+        if !merged_signatures.is_enough() {
+            eprintln!(
+                "still require {} signatures",
+                merged_signatures.threshold() as usize - merged_signatures.signatures().len()
+            );
+        } else {
+            eprintln!("enough signatures collected for the multisig txn, txn can be submitted now");
+        }
+
+        // construct the signed txn with merged signatures.
+        let signed_txn = {
+            let authenticator = TransactionAuthenticator::MultiEd25519 {
+                public_key: multisig_public_key,
+                signature: merged_signatures.into(),
+            };
+            SignedUserTransaction::new(partial_signed_txn.into_raw_transaction(), authenticator)
+        };
+
+        if submit {
+            let _ = self.submit_txn(signed_txn, blocking);
+            return Ok(PathBuf::new());
+        }
+
+        // output the txn, send this to other participants to sign, or just submit it.
+        let output_file = {
+            let mut output_dir = output_dir;
+            // use hash's as output file name
+            let file_name = signed_txn.crypto_hash().to_hex();
+            output_dir.push(file_name);
+            output_dir.set_extension("multisig-txn");
+            output_dir
+        };
+        let mut file = File::create(output_file.clone())?;
+        // write txn to file
+        bcs_ext::serialize_into(&mut file, &signed_txn)?;
+        Ok(output_file)
     }
 
     pub fn submit_txn(
